+++ EventStarterApp/app/src/main/java/com/br/italoscompany/eventstarterapp/Functionalities/UserPerfil/UserPerfilPresenter.java
@@ -16,7 +16,7 @@ public class UserPerfilPresenter implements IUserPerfil.IPresenter {
     }
 
     @Override
-    public void onShowUser(long userId) {
+    public void onShowUser(int userId) {
         userIView.showUser(this.userModel.findUserById(userId));
     }
 }
